Color Bar Graph with Tooltip
