--- src/app/modules/d3/components/d3-away-container/d3-away/d3-away.component.scss
@@ -1,7 +1,0 @@
-.bar {
-    fill: steelblue;
-  }
-  
-  .bar:hover {
-    fill: brown;
-  }
