+++ src/app/modules/d3/components/d3-color-bar-container/d3-color-bar-container.component.spec.ts
@@ -1,20 +1,20 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { D3AwayContainerComponent } from './d3-away-container.component';
+import { D3ColorBarContainerComponent } from './d3-color-bar-container.component';
 
 describe('D3AwayContainerComponent', () => {
-  let component: D3AwayContainerComponent;
-  let fixture: ComponentFixture<D3AwayContainerComponent>;
+  let component: D3ColorBarContainerComponent;
+  let fixture: ComponentFixture<D3ColorBarContainerComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ D3AwayContainerComponent ]
+      declarations: [D3ColorBarContainerComponent]
     })
-    .compileComponents();
+      .compileComponents();
   });
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(D3AwayContainerComponent);
+    fixture = TestBed.createComponent(D3ColorBarContainerComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
