+++ src/app/modules/d3/components/d3-color-bar-container/d3-color-bar-container.component.ts
@@ -3,10 +3,10 @@ import { Observable } from 'rxjs';
 import { HomeService } from 'src/app/services/home.service';
 
 @Component({
-  selector: 'app-d3-away-container',
-  template: `<app-d3-away    [barChartData] = "barChartData$  | async"> </app-d3-away>`
+  selector: 'app-d3-color-bar-container',
+  template: `<app-d3-color-bar    [barChartData] = "barChartData$  | async"> </app-d3-color-bar>`
 })
-export class D3AwayContainerComponent implements OnInit {
+export class D3ColorBarContainerComponent implements OnInit {
 
   barChartData$: Observable<any>
