+++ src/app/modules/d3/components/d3-color-bar-container/d3-color-bar/d3-color-bar.component.html
@@ -3,7 +3,7 @@
     <div class="row">
       <ng-container *ngFor="let _ of barChartData.data">
         <button class="m-3"
-                (click)="firstDraft(_)">{{_.fileName}}</button><br>
+                (click)="triggerGraph(_)">{{_.fileName}}</button><br>
       </ng-container>
     </div>
     <div class="row">
