+++ src/app/modules/d3/components/d3-color-bar-container/d3-color-bar/d3-color-bar.component.scss
@@ -1,0 +1,11 @@
+.bar {
+  fill: steelblue;
+}
+
+.bar:hover {
+  fill: brown;
+}
+
+.rect {
+  padding: 0.2rem;
+}
