+++ src/app/modules/d3/components/d3-shell/d3-shell.component.html
@@ -1,7 +1,6 @@
 <div class="container h-100">
   <nav class="navbar navbar-dark bg-dark">
-    <!-- <span class="navbar-brand h1">{{ (chartLibrary$ | async)?.name }}</span> -->
     <span class="navbar-brand h1">{{ chartLibrary.name }}</span>
   </nav>
-  <app-d3-away-container></app-d3-away-container>
+  <router-outlet></router-outlet>
 </div>
